+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - [#1869] Replace references now replaces all references, not just PlaceObject
 - Handle StartSound tag as CharacterIdTag
+- Clearing shape export cache on changes
 
 ## [16.3.1] - 2022-11-14
 ### Fixed